requirements.txt
nibabel==5.2.0
numpy==1.24.4
openpyxl==3.1.2
Pillow==10.2.0
PySimpleGUI==4.60.5
scipy==1.12.0
